pyproject.toml
[project]
name = "agentic-tdd"
version = "0.1.0"
description = "Agentic TDD 2.5-Loop System"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "radon>=6.0",
    "rich>=13.0",
]

[project.scripts]
atdd = "orchestrator.__main__:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "hypothesis>=6.0",
    "pytest-timeout>=2.0",
    "pytest-benchmark>=4.0",
]
docs = [
    "mkdocs-material>=9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["orchestrator"]

[tool.pytest.ini_options]
